Optimized Makefile for continuation. Added support for manual entries.
Makefile is now structured by files such that it can optimally determine
whether or not to redo a certain target. It takes advantage of the
'checksum' files to determine whether or not to do so.

Introduced 'pgp-keys-manual.txt' for manual entries to add to the final
'pgp-keys.map' file. Renamed 'pgp-keys-raw.txt' to
'pgp-keys-generated.txt' to better reflect its origin.